add SimpleTimer 라이브 액티비티
