+++ WOD/Timer.xcodeproj/xcuserdata/kyungheonlee.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,12 +7,25 @@
 		<key>Timer.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
 		</dict>
 		<key>TimerWidgetExtension.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>56AF0AD42AE892B5009AF8A2</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>BE62D1512B312C2400E1B233</key>
+		<dict>
+			<key>primary</key>
+			<true/>
 		</dict>
 	</dict>
 </dict>
