--- WOD/Timer/Common/Controller/Widget/WidgetManager.swift
@@ -1,13 +1,0 @@
-//
-//  WidgetManager.swift
-//  Timer
-//
-//  Created by KyungHeon Lee on 2023/12/19.
-//
-
-import Foundation
-
-class WidgetManager: ObservableObject {
-    @Published var contentState = TimerWidgetAttributes.ContentState(currentDisplayTime: 0)
-    
-}
