+++ WOD/TimerWidget/TimerWidgetAttributes.swift
@@ -1,0 +1,21 @@
+//
+//  TimerWidgetAttributes.swift
+//  Timer
+//
+//  Created by KyungHeon Lee on 2023/12/19.
+//
+
+import ActivityKit
+import WidgetKit
+import SwiftUI
+
+// MARK: - SimpleWidgetAttributes
+struct SimpleWidgetAttributes: ActivityAttributes {
+    public struct ContentState: Codable, Hashable {
+        var currentState: String
+        var currentRound: Int
+        var currentDisplayTime: Int
+    }
+    
+    var name: String
+}
